+++ packages/mold/package.json
@@ -18,7 +18,7 @@
     "d3": "^7.8.5",
     "eslint": "8.46.0",
     "eslint-config-next": "13.4.12",
-    "next": "13.4.12",
+    "next": "canary",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-redux": "^8.1.2",
