+++ packages/mold/src/app/components/editor/index.tsx
@@ -1,0 +1,13 @@
+'use-client'
+
+import GridControls from '@/app/components/grid-controls'
+
+const Editor = () => {
+  return (
+    <section>
+      <GridControls />
+    </section>
+  )
+}
+
+export default Editor
