+++ packages/mold/src/app/components/notLargeEnough/NotLargeEnough.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
-import NotLargeEnough from './NotLargeEnough'
+import NotLargeEnough from '.'
 
 describe('NotLargeEnough component', () => {
   test('renders correct message in the component', () => {
