+++ packages/mold/src/app/components/wrapper/wrapper.module.css
@@ -1,0 +1,5 @@
+.svgCanvas {
+  display: flex;
+  justify-content: space-between;
+  flex-direction: row;
+}
