+++ packages/mold/src/app/page.tsx
@@ -1,11 +1,16 @@
 import Image from 'next/image'
 import styles from './page.module.css'
-import SvgCanvas from './components/svg-canvas/SvgCanvas'
+import Wrapper from './components/wrapper'
+import Editor from './components/editor'
+import GridControls from './components/grid-controls'
+// import Wrapper from './components/canvas'
 
 export default function Home() {
   return (
     <main className={styles.main}>
-      <SvgCanvas requiredScreenWidth={595} />
+      {/* <Wrapper /> */}
+      {/* <GridControls /> */}
+      <Wrapper requiredScreenWidth={595} />
     </main>
   )
 }
